pyproject.toml
[tool.poetry]
name = "istore-django-project"
version = "0.1.0"
description = ""
authors = ["Sergei Gorlanov <itserg90@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
ipython = "^8.24.0"
pytils = "^0.4.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
django = "4.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
